Move isignal module to core/isignal
