+++ pyanaconda/core/isignal.py
