+++ pyanaconda/input_checking.py
@@ -20,7 +20,7 @@
 
 import pwquality
 
-from pyanaconda.isignal import Signal
+from pyanaconda.core.isignal import Signal
 from pyanaconda.core.i18n import _
 from pyanaconda.core import constants, regexes
 from pyanaconda import users
