+++ pyanaconda/lifecycle.py
@@ -19,7 +19,7 @@
 #
 
 from threading import RLock
-from pyanaconda.isignal import Signal
+from pyanaconda.core.isignal import Signal
 from pyanaconda.core.iutil import synchronized
 
 from pyanaconda.anaconda_loggers import get_module_logger
