+++ pyanaconda/task/task.py
@@ -24,7 +24,7 @@
 from threading import Lock
 from abc import ABC, abstractmethod
 
-from pyanaconda.isignal import Signal
+from pyanaconda.core.isignal import Signal
 from pyanaconda.core.constants import THREAD_DBUS_TASK
 from pyanaconda.task.task_interface import TaskAlreadyRunningException
 from pyanaconda.threading import threadMgr, AnacondaThread
