+++ pyanaconda/ui/common.py
@@ -22,7 +22,7 @@ from abc import ABCMeta, abstractproperty
 from pyanaconda.core.constants import ANACONDA_ENVIRON, FIRSTBOOT_ENVIRON
 from pyanaconda import screen_access
 from pyanaconda.core.iutil import collect
-from pyanaconda.isignal import Signal
+from pyanaconda.core.isignal import Signal
 from pyanaconda import lifecycle
 
 from pykickstart.constants import FIRSTBOOT_RECONFIG, DISPLAY_MODE_TEXT
