+++ tests/pyanaconda_tests/signal_test.py
@@ -22,7 +22,7 @@
 
 import unittest
 
-from pyanaconda.isignal import Signal
+from pyanaconda.core.isignal import Signal
 
 class FooClass(object):
     def __init__(self):
